Also fix the active player popover
